feat: 维兹南灵魂瓦解与无尽模式相关改动
- eb_veznan：Soul Drain、spellDown 时长与朝向、封印/传送门逻辑；灵魂打断冷却

- 无尽：endless 系统拆分、塔技能协议、科技注册与试点

- 音效/文案与资源索引、game_animations 等配套更新

Made-with: Cursor
