+++ _assets/kr1-desktop/strings/zh-Hans.lua
@@ -6959,5 +6959,47 @@ PLEASE DON'T"] = "单击此处以跳过。\n请不要",
 	barrack_dominant_NAME = "绝对统率",
 	barrack_dominant_DESCRIPTION = "大大提高士兵的调遣范围，并提高士兵的移动速度。",
 	mage_treasure_NAME = "奥术结晶",
-	mage_treasure_DESCRIPTION = "被法球命中的敌人赏金提高，最高提高10%。"
+	mage_treasure_DESCRIPTION = "被法球命中的敌人赏金提高，最高提高10%。",
+	TOWER_SKILL_BARREL_FIELD_RATION_NAME = "战地补给",
+	TOWER_SKILL_BARREL_FIELD_RATION_DESC = "酒桶塔周期性治疗附近友军士兵。",
+	TOWER_SKILL_ARCANE_OVERLOAD_NAME = "奥术过载",
+	TOWER_SKILL_ARCANE_OVERLOAD_DESC = "奥术塔周期性释放脉冲打击范围敌人。",
+	TOWER_SKILL_BALLISTA_SALVO_NAME = "弩炮齐射",
+	TOWER_SKILL_BALLISTA_SALVO_DESC = "弩炮塔周期性对多个敌人执行强力齐射。",
+	TOWER_SKILL_PALADIN_BATTLE_PRAYER_NAME = "战阵祷告",
+	TOWER_SKILL_PALADIN_BATTLE_PRAYER_DESC = "圣骑士塔周期性治疗范围内友军单位。",
+	TOWER_SKILL_STARGAZER_COMET_BURST_NAME = "彗光迸发",
+	TOWER_SKILL_STARGAZER_COMET_BURST_DESC = "星相塔周期性对范围敌人降下彗光打击。",
+	TOWER_SKILL_FLAMESPITTER_BURST_NAME = "喷焰超压",
+	TOWER_SKILL_FLAMESPITTER_BURST_DESC = "喷火塔周期性强化喷焰，对多个敌人造成额外伤害。",
+	TOWER_SKILL_DARK_ELF_EXECUTION_ORDER_NAME = "处决号令",
+	TOWER_SKILL_DARK_ELF_EXECUTION_ORDER_DESC = "暗精灵塔周期性对多个目标发动处决射击。",
+	TOWER_SKILL_TRICANNON_SHOCK_BARRAGE_NAME = "三炮震荡连射",
+	TOWER_SKILL_TRICANNON_SHOCK_BARRAGE_DESC = "三管炮塔周期性发起震荡齐射，重创前线敌人。",
+	TOWER_SKILL_RAY_CHAIN_OVERLOAD_NAME = "连锁过载",
+	TOWER_SKILL_RAY_CHAIN_OVERLOAD_DESC = "射线塔周期性引发连锁过载，打击范围内多名敌人。",
+	TOWER_SKILL_DWARF_BATTLE_ALE_NAME = "战酿鼓舞",
+	TOWER_SKILL_DWARF_BATTLE_ALE_DESC = "矮人塔周期性为附近友军恢复生命。",
+	TOWER_SKILL_GHOST_REAP_ORDER_NAME = "幽镰收割令",
+	TOWER_SKILL_GHOST_REAP_ORDER_DESC = "幽魂塔周期性释放收割指令，对多个敌人造成重创。",
+	TOWER_SKILL_NECROMANCER_DEATH_CHANT_NAME = "亡咏咒令",
+	TOWER_SKILL_NECROMANCER_DEATH_CHANT_DESC = "死灵塔周期性释放亡咏，对范围敌人施加额外打击。",
+	TOWER_SKILL_TOAD_TOXIC_BURST_NAME = "剧毒喷发",
+	TOWER_SKILL_TOAD_TOXIC_BURST_DESC = "蟾蜍塔周期性喷发毒击，压制附近敌群。",
+	TOWER_SKILL_DEMON_PIT_HELL_PRESSURE_NAME = "炼狱压迫",
+	TOWER_SKILL_DEMON_PIT_HELL_PRESSURE_DESC = "恶魔坑塔周期性释放炼狱能量冲击敌人。",
+	TOWER_SKILL_ROCKET_GUNNERS_SATURATION_FIRE_NAME = "饱和火力",
+	TOWER_SKILL_ROCKET_GUNNERS_SATURATION_FIRE_DESC = "火箭枪手塔周期性进行范围压制射击。",
+	TOWER_SKILL_SANDSTORM_FOCUS_FIRE_NAME = "风沙点杀",
+	TOWER_SKILL_SANDSTORM_FOCUS_FIRE_DESC = "沙丘塔周期性发起风沙点杀，压制前线目标。",
+	TOWER_SKILL_ROYAL_ARCHERS_HAIL_ORDER_NAME = "王室箭雨令",
+	TOWER_SKILL_ROYAL_ARCHERS_HAIL_ORDER_DESC = "皇家弓箭手塔周期性下达齐射命令，覆盖敌群。",
+	TOWER_SKILL_PANDA_STORM_COMBO_NAME = "熊猫连携风暴",
+	TOWER_SKILL_PANDA_STORM_COMBO_DESC = "熊猫塔周期性打出连携风暴，对敌群造成额外伤害。",
+	TOWER_SKILL_GEODE_OVERCHARGE_PULSE_NAME = "晶核过载脉冲",
+	TOWER_SKILL_GEODE_OVERCHARGE_PULSE_DESC = "晶核塔周期性释放过载脉冲，打击范围敌人。",
+	TOWER_SKILL_ARBOREAN_NATURE_SURGE_NAME = "林灵涌动",
+	TOWER_SKILL_ARBOREAN_NATURE_SURGE_DESC = "林灵使塔周期性引导自然涌动，打击附近敌人。",
+	TOWER_SKILL_DRAGONS_SKY_RUIN_NAME = "苍穹崩灭",
+	TOWER_SKILL_DRAGONS_SKY_RUIN_DESC = "巨龙塔周期性发动空域毁灭打击，覆盖敌群。"
 }
