+++ all/game.lua
@@ -101,6 +101,7 @@ game.simulation_systems = {
 	"pops",
 	"goal_line",
 	"tower_upgrade",
+	"tower_skill",
 	"game_upgrades",
 	"texts",
 	"particle_system",
