+++ all/systems.lua
@@ -4,6 +4,7 @@ local log = require("lib.klua.log"):new("systems")
 local km = require("lib.klua.macros")
 local signal = require("lib.hump.signal")
 local perf = require("dove_modules.perf.perf")
+local SystemsIndex = require("systems.index")
 local IS_ANDROID = love.system.getOS() == "Android"
 require("lib.klua.table")
 require("lib.klua.dump")
@@ -396,1554 +397,6 @@ function sys.level:on_update(dt, ts, store)
 	perf.stop("level")
 end
 
-sys.events = {}
-sys.events.name = "events"
-
-function sys.events:init(store)
-	store.event_handlers = {}
-end
-
-function sys.events:on_insert(entity, store)
-	if entity.events then
-		for _, ev in pairs(entity.events.list) do
-			if not store.event_handlers[ev.name] then
-				store.event_handlers[ev.name] = {}
-			end
-
-			ev.entity_id = entity.id
-
-			table.insert(store.event_handlers[ev.name], ev)
-		end
-	end
-
-	return true
-end
-
-function sys.events:on_remove(entity, store)
-	if entity.events then
-		for _, ev in pairs(entity.events.list) do
-			if store.event_handlers[ev.name] then
-				table.removeobject(store.event_handlers[ev.name], ev)
-			end
-		end
-	end
-
-	return true
-end
-
-sys.wave_spawn_tsv = {}
-sys.wave_spawn_tsv.name = "wave_spawn_tsv"
-sys.wave_spawn_tsv.cmd_fns = {}
-
-function sys.wave_spawn_tsv.cmd_fns.column_names(store, cmd)
-	if cmd.time_columns then
-		W.db.time_columns = table.deepclone(cmd.time_columns)
-	end
-
-	if cmd.path_columns then
-		W.db.path_columns = table.deepclone(cmd.path_columns)
-	end
-end
-
-function sys.wave_spawn_tsv.cmd_fns.flags(store, cmd)
-	if not cmd.flags_visibility then
-		return
-	end
-
-	W.db.flags_visibility = cmd.flags_visibility
-end
-
-function sys.wave_spawn_tsv.cmd_fns.manual_wave(store, cmd)
-	log.debug("manual_wave started: %s", cmd.wave_name)
-end
-
-function sys.wave_spawn_tsv.cmd_fns.manual_wave_repeat(store, cmd)
-	local mws = W:get_wave_status(cmd.wave_name)
-
-	if not mws then
-		log.error("manual_wave_repeat: manual_wave_index[%s] not found", cmd.wave_name)
-
-		return
-	end
-
-	if mws.repeat_remaining == -1 then
-		mws.current_idx = mws.first_idx
-	elseif mws.repeat_remaining and mws.repeat_remaining > 0 then
-		mws.current_idx = mws.first_idx
-		mws.repeat_remaining = mws.repeat_remaining - 1
-	else
-		mws.status = W.WS_DONE
-	end
-end
-
-function sys.wave_spawn_tsv.cmd_fns.call_manual_wave(store, cmd)
-	log.debug("call_manual_wave: %s", cmd.value)
-	W:start_manual_wave(cmd.value)
-end
-
-function sys.wave_spawn_tsv.cmd_fns.wave(store, cmd)
-	local group = W:create_wave_group_from_tsv(cmd)
-
-	group.group_idx = store.wave_group_number + 1
-	store.next_wave_group_ready = group
-
-	signal.emit("next-wave-ready", group)
-
-	local wave_number = store.wave_group_number
-	local wait_time = cmd.wait_time
-	local start_ts = store.tick_ts
-
-	if wait_time < 0 then
-		while not store.send_next_wave and not store.force_next_wave do
-			coroutine.yield()
-		end
-	else
-		U.y_wait(store, wait_time, function(store, wait_time)
-			return store.send_next_wave or store.force_next_wave
-		end)
-
-	end
-
-	local actual_wait_time = store.tick_ts - start_ts
-
-	store.next_wave_group_ready = nil
-	store.wave_group_number = store.wave_group_number + 1
-
-	if store.force_next_wave then
-		store.force_next_wave = false
-	end
-
-	if store.send_next_wave == true and store.wave_group_number > 1 then
-		local score_reward
-		local remaining_secs = km.round(wait_time - actual_wait_time)
-
-		if store.level_mode == GAME_MODE_ENDLESS then
-			store.early_wave_reward = math.ceil(remaining_secs * GS.early_wave_reward_per_second * W:get_endless_early_wave_reward_factor())
-
-			local conf = W:get_endless_score_config()
-			local time_factor = km.clamp(0, 1, remaining_secs / fts(group.interval))
-
-			score_reward = km.round((wave_number - 1) * conf.scorePerWave * conf.scoreNextWaveMultiplier * time_factor * #group.waves)
-			store.player_score = store.player_score + score_reward
-
-			log.debug("ENDLESS: early wave %s reward %s (time_factor:%s scorePerWave:%s scoreNextWaveMultiplier:%s flags:%s", wave_number, score_reward, time_factor, conf.scorePerWave, conf.scoreNextWaveMultiplier, #group.waves)
-		else
-			store.early_wave_reward = math.ceil(remaining_secs * GS.early_wave_reward_per_second)
-		end
-
-		store.player_gold = store.player_gold + store.early_wave_reward
-
-		signal.emit("early-wave-called", group, store.early_wave_reward, remaining_secs, score_reward)
-	else
-		store.early_wave_reward = 0
-	end
-
-	if store.level_mode == GAME_MODE_ENDLESS and wave_number > 1 then
-		local conf = W:get_endless_score_config()
-		local reward = (wave_number - 1) * conf.scorePerWave
-
-		store.player_score = store.player_score + reward
-	end
-
-	store.current_spawn_idx = 0
-	store.send_next_wave = false
-	store.current_wave_group = group
-
-	signal.emit("next-wave-sent", group)
-end
-
-function sys.wave_spawn_tsv.cmd_fns.spawn(store, cmd, wave_name)
-	local wait_time = cmd.wait_time
-	local multiplier = store.config.enemy_count_multiplier
-
-	for count = 1, multiplier do
-		if wait_time and wait_time > 0 then
-			U.y_wait(store, wait_time / multiplier, function(store, wait_time)
-				return store.force_next_wave
-			end)
-		end
-
-		if store.force_next_wave then
-			log.debug("skipping spawn command due to force_next_wave")
-
-			return
-		end
-
-		for _, o in pairs(cmd.spawns) do
-			if not U.is_seen(store, o.enemy) then
-				signal.emit("wave-notification", "icon", o.enemy)
-				U.mark_seen(store, o.enemy)
-			end
-
-			local e = E:create_entity(o.enemy)
-
-			if e then
-				store.current_spawn_idx = store.current_spawn_idx + 1
-
-				local path = P.paths[o.pi]
-
-				e.nav_path.pi = o.pi
-				e.nav_path.spi = o.spi == "*" and math.random(#path) or o.spi
-				e.nav_path.ni = P:get_start_node(o.pi)
-
-				queue_insert(store, e)
-			else
-				log.error("Entity template %s not found", o.enemy)
-			end
-		end
-	end
-end
-
-function sys.wave_spawn_tsv.cmd_fns.event(store, cmd)
-	local wait_time = cmd.wait_time
-
-	if wait_time and wait_time > 0 then
-		U.y_wait(store, wait_time, function(store, wait_time)
-			return store.force_next_wave
-		end)
-	end
-
-	local handlers = store.event_handlers
-
-	if cmd.event_name and handlers and handlers[cmd.event_name] then
-		for _, ev in pairs(handlers[cmd.event_name]) do
-			local entity = store.entities[ev.entity_id]
-
-			ev.on_event(entity, store, ev.name, unpack(cmd.event_params or {}))
-		end
-	end
-end
-
-function sys.wave_spawn_tsv.cmd_fns.signal(store, cmd)
-	local wait_time = cmd.wait_time
-
-	if wait_time and wait_time > 0 then
-		U.y_wait(store, wait_time, function(store, wait_time)
-			return store.force_next_wave
-		end)
-	end
-
-	if cmd.signal_name then
-		signal.emit(cmd.signal_name, unpack(cmd.signal_params or {}))
-	end
-end
-
-function sys.wave_spawn_tsv.cmd_fns.wait_signal(store, cmd)
-	local signal_name = cmd.signal_name
-
-	store.wait_signal_done = nil
-
-	local function fn(...)
-		log.debug("wait_signal : signal received")
-
-		store.wait_signal_done = "arrived"
-	end
-
-	if signal_name then
-		log.debug("wait_signal: registering signal %s", signal_name)
-		signal.register(signal_name, fn)
-	end
-
-	log.debug("wait_signal: waiting for signal:%s  time:%s", signal_name, cmd.wait_time)
-
-	--if cmd.wait_time < 0 then
-	if cmd.wait_time < 0 then -- or signal_name == "all_enemies_died" or signal_name == "ciclone_end" or signal_name == "barrage_end" then
-		while not store.wait_signal_done and not store.force_next_wave do
-			coroutine.yield()
-		end
-	else
-		--新clamp 20251003
-		cmd.wait_time = km.clamp(30, 60, cmd.wait_time)
-		U.y_wait(store, cmd.wait_time, function(store, wait_time)
-			if store.wait_signal_done or store.force_next_wave then
-				store.wait_signal_done = "interrupted"
-
-				return true
-			end
-		end)
-	end
-
-	log.debug("wait_signal: deregistering signal %s", signal_name)
-	signal.remove(signal_name, fn)
-end
-
-function sys.wave_spawn_tsv.cmd_fns.wait(store, cmd)
-	local wait_time = cmd.wait_time
-
-	if wait_time and wait_time > 0 then
-		U.y_wait(store, wait_time, function(store, wait_time)
-			return store.force_next_wave
-		end)
-	end
-end
-
-function sys.wave_spawn_tsv.y_run_wave(store, wave_name)
-	local cmd_fns = sys.wave_spawn_tsv.cmd_fns
-	local cmd = W:get_next_cmd(wave_name)
-
-	while cmd do
-		if cmd_fns[cmd.name] then
-			cmd_fns[cmd.name](store, cmd, wave_name)
-		elseif cmd.wait_time then
-			if cmd.wait_time < 0 then
-				while not store.send_next_wave and not store.force_next_wave do
-					coroutine.yield()
-				end
-			else
-				U.y_wait(store, cmd.wait_time, function(store, wait_time)
-					return store.force_next_wave
-				end)
-			end
-		end
-
-		cmd = W:get_next_cmd(wave_name)
-	end
-
-	return true
-end
-
-function sys.wave_spawn_tsv:init(store)
-	if W.format ~= "tsv" then
-		return "skip"
-	end
-
-	store.wave_group_number = 0
-	store.waves_finished = false
-	store.last_wave_ts = 0
-	store.send_next_wave = false
-	store.manual_wave_cos = {}
-
-	do
-		local cmd_fns = sys.wave_spawn_tsv.cmd_fns
-		local cmd = W:peek_next_cmd()
-
-		while cmd and cmd.name ~= "wave" and cmd.name ~= "manual_wave" and not cmd.wait_time do
-			if cmd_fns[cmd.name] then
-				cmd_fns[cmd.name](store, cmd)
-			end
-
-			cmd = W:get_next_cmd()
-			cmd = W:peek_next_cmd()
-		end
-	end
-
-	if store.level_mode == GAME_MODE_ENDLESS then
-		store.wave_group_total = 0
-	else
-		store.wave_group_total = W:groups_count()
-	end
-
-	store.wave_spawn_co = coroutine.create(sys.wave_spawn_tsv.y_run_wave)
-end
-
-function sys.wave_spawn_tsv:on_update(dt, ts, store)
-	if store.force_next_wave then
-		LU.kill_all_enemies(store, nil, true)
-	end
-
-	if store.wave_spawn_co then
-		local ok, done = coroutine.resume(store.wave_spawn_co, store)
-
-		if ok and done then
-			store.wave_spawn_co = nil
-			store.waves_finished = true
-
-			log.debug("wave_spawn_tsv: waves finished")
-		end
-
-		if not ok then
-			log.error("wave_spawn_tsv: Error resuming wave_spawn_co co:%s", debug.traceback(store.wave_spawn_co, done))
-
-			store.wave_spawn_co = nil
-		end
-	elseif store.waves_finished == false then
-		--流辉 20251003 否则，当没有线程的时候，强制结束
-		local level_table = {116}
-		if store.wave_group_total <= (store.wave_group_number + 0.5) and not table.contains(level_table, store.level_idx) then
-			store.waves_finished = true
-			print("LIUHUI349 FORCE END LEVEL")
-		end
-	end
-
-	if W:has_pending_manual_waves() then
-		for _, name in pairs(W:list_pending_manual_waves()) do
-			local ws = W:get_wave_status(name)
-
-			ws.state = W.WS_RUNNING
-			store.manual_wave_cos = store.manual_wave_cos or {}
-			store.manual_wave_cos[name] = coroutine.create(sys.wave_spawn_tsv.y_run_wave)
-		end
-	end
-
-	if store.manual_wave_cos then
-		local to_remove
-
-		for name, co in pairs(store.manual_wave_cos) do
-			local ws = W:get_wave_status(name)
-
-			if ws and ws.state == W.WS_DONE then
-				to_remove = to_remove or {}
-
-				table.insert(to_remove, name)
-			else
-				local ok, done = coroutine.resume(co, store, name)
-
-				if ok and done then
-					to_remove = to_remove or {}
-
-					table.insert(to_remove, name)
-				end
-
-				if not ok then
-					log.error("wave_spawn_tsv: Error resuming manual_wave_cos[%s]:%s", name, debug.traceback(co, done))
-
-					store.wave_spawn_co = nil
-				end
-			end
-		end
-
-		if to_remove then
-			for _, name in pairs(to_remove) do
-				store.manual_wave_cos[name] = nil
-
-				local ws = W:get_wave_status(name)
-
-				if ws then
-					ws.state = W.WS_REMOVED
-				end
-			end
-
-			to_remove = nil
-		end
-	end
-
-	store.force_next_wave = false
-end
-
-sys.wave_spawn = {}
-sys.wave_spawn.name = "wave_spawn"
-
-local function spawner(store, wave, group_id)
-	log.debug("spawner thread(%s) for wave(%s) starting", coroutine.running(), tostring(wave))
-
-	local spawns = wave.spawns
-	local pi = wave.path_index
-	local last_spawn_ts = 0
-
-	for i = 1, #spawns do
-		for count = 1, store.config.enemy_count_multiplier do
-			local current_count = 0
-			local current_creep
-			local s = spawns[i]
-			local path = P.paths[pi]
-
-			if not U.is_seen(store, s.creep) then
-				signal.emit("wave-notification", "icon", s.creep)
-				U.mark_seen(store, s.creep)
-			end
-
-			if s.creep_aux and not U.is_seen(store, s.creep_aux) then
-				signal.emit("wave-notification", "icon", s.creep_aux)
-				U.mark_seen(store, s.creep_aux)
-			end
-
-			for j = 1, s.max do
-				U.y_wait(store, fts(s.interval or 0) / store.config.enemy_count_multiplier)
-
-				if not current_creep then
-					current_creep = s.creep
-				elseif s.creep_aux and s.max_same and s.max_same > 0 and current_count >= s.max_same then
-					current_creep = s.creep == current_creep and s.creep_aux or s.creep
-					current_count = 0
-				end
-
-				local e = E:create_entity(current_creep)
-
-				if e then
-					e.nav_path.pi = pi
-					e.nav_path.spi = s.fixed_sub_path == 1 and s.path or random(#path)
-					e.nav_path.ni = P:get_start_node(pi)
-					e.spawn_data = s.spawn_data
-
-					queue_insert(store, e)
-
-					current_count = current_count + 1
-				else
-					log.error("Entity template not found for %s.", s.crep)
-				end
-			end
-
-			if s.max == 0 then
-				U.y_wait(store, fts(s.interval or 0) / store.config.enemy_count_multiplier)
-			end
-
-			local oes = s.on_end_signal
-
-			if oes then
-				log.info("Sending spawner on_end_signal: %s", oes)
-
-				store.wave_signals[oes] = {}
-			end
-
-			if i < #spawns then
-				local interval_next = s.interval_next or 0
-
-				if DI.level == DIFFICULTY_HARD then
-					if group_id > 12 then
-						store.last_wave_ts = store.last_wave_ts - interval_next * 0.75
-						interval_next = interval_next * 0.25
-					elseif group_id > 9 then
-						store.last_wave_ts = store.last_wave_ts - interval_next * 0.5
-						interval_next = interval_next * 0.5
-					elseif group_id > 6 then
-						store.last_wave_ts = store.last_wave_ts - interval_next * 0.25
-						interval_next = interval_next * 0.75
-					end
-				end
-
-				U.y_wait(store, fts(interval_next) / store.config.enemy_count_multiplier)
-			end
-		end
-	end
-
-	log.debug("spawner thread(%s) for wave(%s) about to finish", coroutine.running(), tostring(wave))
-
-	return true
-end
-
-function sys.wave_spawn:init(store)
-	if W.format ~= "lua" then
-		return "skip"
-	end
-	store.wave_group_number = 0
-	store.waves_finished = false
-	store.last_wave_ts = 0
-	store.waves_active = {}
-	store.wave_signals = {}
-	store.send_next_wave = false
-
-	if store.level_mode_override == GAME_MODE_ENDLESS then
-		store.wave_group_total = 0
-
-		if store.endless and store.endless.wave_group_number then
-			store.wave_group_number = store.endless.wave_group_number
-		end
-	else
-		store.wave_group_total = W:groups_count()
-	end
-
-	local function run(store)
-		log.info("Wave group spawn thread STARTING")
-
-		local i = 1
-		local start = true
-
-		if store.endless and store.endless.wave_group_number then
-			i = store.endless.wave_group_number
-		end
-
-		while W:has_group(i) do
-			local group = W:get_group(i)
-
-			group.group_idx = i
-			store.next_wave_group_ready = group
-
-			signal.emit("next-wave-ready", group)
-
-			if start then
-				group.group_idx = 1
-
-				for _, wave in pairs(group.waves) do
-					if wave.notification and wave.notification ~= "" then
-						signal.emit("wave-notification", "view", wave.notification)
-					end
-				end
-
-				while not store.send_next_wave do
-					coroutine.yield()
-				end
-
-				start = false
-
-				log.debug("Sending first WAVE. (Started by player)")
-			else
-				while not store.send_next_wave and not (store.tick_ts - store.last_wave_ts >= fts(group.interval)) and not store.force_next_wave do
-					coroutine.yield()
-				end
-			end
-
-			log.info("sending WAVE group %02d (%02d waves)", i, #group.waves)
-
-			store.next_wave_group_ready = nil
-			store.wave_group_number = i
-
-			if store.send_next_wave == true and i > 1 then
-				local score_reward
-				local remaining_secs = km.round(fts(group.interval) - (store.tick_ts - store.last_wave_ts))
-
-				if store.level_mode == -1 then
-					-- if store.level_mode == GAME_MODE_ENDLESS then
-					store.early_wave_reward = ceil(remaining_secs * GS.early_wave_reward_per_second * W:get_endless_early_wave_reward_factor())
-
-					local conf = W:get_endless_score_config()
-					local time_factor = km.clamp(0, 1, remaining_secs / fts(group.interval))
-
-					score_reward = km.round((i - 1) * conf.scorePerWave * conf.scoreNextWaveMultiplier * time_factor * #group.waves)
-					store.player_score = store.player_score + score_reward
-
-					log.debug("ENDLESS: early wave %s reward %s (time_factor:%s scorePerWave:%s scoreNextWaveMultiplier:%s flags:%s", i, score_reward, time_factor, conf.scorePerWave, conf.scoreNextWaveMultiplier, #group.waves)
-				else
-					store.early_wave_reward = ceil(remaining_secs * GS.early_wave_reward_per_second)
-				end
-
-				store.player_gold = store.player_gold + store.early_wave_reward
-
-				signal.emit("early-wave-called", group, store.early_wave_reward, remaining_secs, score_reward)
-			else
-				store.early_wave_reward = 0
-
-				if store.criket then
-					store.criket.start_time = store.tick_ts
-				end
-			end
-
-			store.send_next_wave = false
-			store.current_wave_group = group
-
-			signal.emit("next-wave-sent", group)
-
-			for j, wave in pairs(group.waves) do
-				wave.group_idx = i
-
-				if i ~= 1 and wave.notification and wave.notification ~= "" then
-					signal.emit("wave-notification", "view", wave.notification)
-				end
-
-				if wave.notification_second_level and wave.notification_second_level ~= "" then
-					signal.emit("wave-notification", "icon", wave.notification_second_level)
-				end
-
-				local sco = coroutine.create(function()
-					local wave_start_ts = store.tick_ts
-
-					while store.tick_ts < wave_start_ts + fts(wave.delay) do
-						coroutine.yield()
-					end
-
-					return spawner(store, wave, i)
-				end)
-
-				store.waves_active[sco] = sco
-			end
-
-			log.info("WAVE group %d about to wait for all its spawner threads to finish", i)
-
-			while next(store.waves_active) do
-				coroutine.yield()
-			end
-
-			store.current_wave_group = nil
-			store.last_wave_ts = store.tick_ts
-			i = i + 1
-		end
-
-		log.info("WAVE spawn thread FINISHED")
-
-		return true
-	end
-
-	store.wave_spawn_thread = coroutine.create(run)
-end
-
-function sys.wave_spawn:force_next_wave(store)
-	if store.force_next_wave then
-		store.waves_active = {}
-
-		LU.kill_all_enemies(store, nil, true)
-	end
-end
-
-function sys.wave_spawn:on_update(dt, ts, store)
-	perf.start("wave_spawn")
-	sys.wave_spawn:force_next_wave(store)
-
-	if store.wave_spawn_thread then
-		local ok, done = coroutine.resume(store.wave_spawn_thread, store)
-
-		if ok and done then
-			store.wave_spawn_thread = nil
-			store.waves_finished = true
-
-			log.debug("++++ WAVES FINISHED")
-		end
-
-		if not ok then
-			log.error("Error resuming wave_spawn_thread co: %s", debug.traceback(store.wave_spawn_thread, done))
-
-			store.wave_spawn_thread = nil
-		end
-	end
-
-	local to_cleanup
-
-	for _, co in pairs(store.waves_active) do
-		local ok, done = coroutine.resume(co, store)
-
-		if ok and done then
-			log.debug("thread (%s) finished after resume()", tostring(co))
-
-			to_cleanup = to_cleanup or {}
-			to_cleanup[#to_cleanup + 1] = co
-		end
-
-		if not ok then
-			local err = done
-
-			log.error("Error resuming spawner thread (%s): %s", tostring(co), debug.traceback(co, err))
-		end
-	end
-
-	if to_cleanup then
-		for _, co in pairs(to_cleanup) do
-			log.debug("removing spawner thread (%s)", co)
-
-			store.waves_active[co] = nil
-		end
-
-		to_cleanup = nil
-	end
-
-	store.force_next_wave = false
-	perf.stop("wave_spawn")
-end
-
-sys.mod_lifecycle = {}
-sys.mod_lifecycle.name = "mod_lifecycle"
-
-function sys.mod_lifecycle:on_insert(entity, store)
-	local mdf = entity.modifier
-
-	if not mdf then
-		return true
-	end
-
-	local this = entity
-	local target_id = mdf.target_id
-	local target = store.entities[target_id]
-
-	if not target then
-		return false
-	end
-
-	if not target._applied_mods then
-		target._applied_mods = {}
-	end
-
-	local modifiers = target._applied_mods
-
-	for i = 1, #modifiers do
-		local m = modifiers[i].modifier
-
-		if m.bans and table.contains(m.bans, this.template_name) then
-			return false
-		end
-	end
-
-	if mdf.remove_banned then
-		for i = 1, #modifiers do
-			local m = modifiers[i]
-			local mm = m.modifier
-
-			if mdf.bans and table.contains(mdf.bans, m.template_name) then
-				mm.removed_by_ban = true
-
-				queue_remove(store, m)
-			end
-
-			if mdf.ban_types and table.contains(mdf.ban_types, mm.type) then
-				mm.removed_by_ban = true
-
-				queue_remove(store, m)
-			end
-		end
-	end
-
-	mdf.ts = store.tick_ts
-
-	if this.render then
-		for i = 1, #this.render.sprites do
-			this.render.sprites[i].ts = store.tick_ts
-		end
-	end
-
-	if mdf.allows_duplicates then
-		return true
-	end
-
-	local duplicates = {}
-
-	for i = 1, #modifiers do
-		local m = modifiers[i]
-
-		if m.template_name == this.template_name then
-			if mdf.level == m.modifier.level and mdf.max_duplicates then
-				mdf.max_duplicates = mdf.max_duplicates - 1
-				duplicates[#duplicates + 1] = m
-
-				if mdf.max_duplicates < 0 then
-					return false
-				end
-			elseif mdf.level > m.modifier.level and mdf.replaces_lower then
-				if m.render then
-					for i = 1, #this.render.sprites do
-						this.render.sprites[i].ts = m.render.sprites[i].ts
-					end
-				end
-
-				queue_remove(store, m)
-			elseif mdf.level == m.modifier.level and mdf.resets_same then
-				m.modifier.ts = store.tick_ts
-
-				if mdf.resets_same_tween and m.tween then
-					m.tween.ts = store.tick_ts - (mdf.resets_same_tween_offset or 0)
-				end
-
-				return false
-			else
-				return false
-			end
-		end
-	end
-
-	if #duplicates > 0 then
-		for _, d in pairs(duplicates) do
-			if d.dps then
-				d.dps.fx = nil
-			end
-
-			if d.render then
-				for i = 1, #d.render.sprites do
-					d.render.sprites[i].hidden = true
-				end
-			end
-		end
-	end
-
-	return true
-end
-
-sys.tower_upgrade = {}
-sys.tower_upgrade.name = "tower_upgrade"
-
-function sys.tower_upgrade:on_update(dt, ts, store)
-	perf.start("tower_upgrade")
-	for _, e in pairs(store.towers) do
-		if e.tower.sell or e.tower.destroy then
-			if e.tower.sell then
-				local refund = store.wave_group_number == 0 and e.tower.spent or km.round(e.tower.refund_factor * e.tower.spent)
-
-				store.player_gold = store.player_gold + refund
-			end
-
-			if e.tower.sell then
-				if e._applied_mods then
-					for _, mod in pairs(e._applied_mods) do
-						queue_remove(store, mod)
-					end
-				end
-			end
-
-			local th = E:create_entity("tower_holder")
-
-			th.pos = V.vclone(e.pos)
-			th.tower.holder_id = e.tower.holder_id
-			th.tower.flip_x = e.tower.flip_x
-
-			if e.tower.default_rally_pos then
-				th.tower.default_rally_pos = e.tower.default_rally_pos
-			end
-
-			if e.tower.terrain_style then
-				th.tower.terrain_style = e.tower.terrain_style
-				th.render.sprites[1].name = string.format(th.render.sprites[1].name, e.tower.terrain_style)
-			end
-
-			if th.ui and e.ui then
-				th.ui.nav_mesh_id = e.ui.nav_mesh_id
-			end
-
-			queue_insert(store, th)
-			queue_remove(store, e)
-			signal.emit("tower-removed", e, th)
-
-			if e.tower.sell then
-				local dust = E:create_entity("fx_tower_sell_dust")
-
-				dust.pos.x, dust.pos.y = th.pos.x, th.pos.y + 35
-				dust.render.sprites[1].ts = store.tick_ts
-
-				queue_insert(store, dust)
-
-				if e.sound_events and e.sound_events.sell then
-					S:queue(e.sound_events.sell, e.sound_events.sell_args)
-				end
-			end
-		elseif e.tower.upgrade_to then
-			if e._applied_mods then
-				for _, mod in pairs(e._applied_mods) do
-					queue_remove(store, mod)
-				end
-			end
-
-			local ne = E:create_entity(e.tower.upgrade_to)
-
-			ne.pos = V.vclone(e.pos)
-			ne.tower.holder_id = e.tower.holder_id
-			ne.tower.flip_x = e.tower.flip_x
-
-			if e.tower.default_rally_pos then
-				ne.tower.default_rally_pos = V.vclone(e.tower.default_rally_pos)
-			end
-
-			if e.tower.terrain_style then
-				ne.tower.terrain_style = e.tower.terrain_style
-				ne.render.sprites[1].name = string.format(ne.render.sprites[1].name, e.tower.terrain_style)
-			end
-
-			if ne.ui and e.ui then
-				ne.ui.nav_mesh_id = e.ui.nav_mesh_id
-			end
-
-			queue_insert(store, ne)
-			queue_remove(store, e)
-			signal.emit("tower-upgraded", ne, e)
-
-			local price = ne.tower.price
-
-			if ne.tower.type == "build_animation" then
-				local bt = E:get_template(ne.build_name)
-
-				price = bt.tower.price
-			elseif e.tower.type == "build_animation" then
-				price = 0
-			elseif e.tower_holder and e.tower_holder.unblock_price > 0 then
-				price = e.tower_holder.unblock_price
-			end
-
-			if e.tower.upgrade_price_multiplier then
-				price = math.ceil(price * e.tower.upgrade_price_multiplier)
-				price = math.floor(price / 10) * 10
-			end
-
-			store.player_gold = store.player_gold - price
-
-			if not e.tower_holder or not e.tower_holder.blocked then
-				ne.tower.spent = e.tower.spent + price
-			end
-
-			if e.tower and e.tower.type == "engineer" and ne.tower.type == "engineer" then
-				if ne.ranged_attack then
-					ne.ranged_attack.ts = e.ranged_attack.ts
-				elseif ne.area_attack then
-					ne.area_attack.ts = e.ranged_attack.ts
-				end
-			elseif e.barrack and ne.barrack then
-				ne.barrack.rally_pos = V.vclone(e.barrack.rally_pos)
-
-				for i, s in ipairs(e.barrack.soldiers) do
-					if s.health.dead then
-					-- block empty
-					else
-						if i > ne.barrack.max_soldiers then
-							U.unblock_target(store, s)
-						else
-							local soldier_type = ne.barrack.soldier_type
-
-							if ne.barrack.soldier_types then
-								soldier_type = ne.barrack.soldier_types[i]
-							end
-
-							local ns = E:create_entity(soldier_type)
-
-							ns.info.i18n_key = s.info.i18n_key
-							ns.soldier.tower_id = ne.id
-							ns.pos = V.vclone(s.pos)
-							ns.motion.dest = V.vclone(s.motion.dest)
-							ns.motion.arrived = s.motion.arrived
-							ns.render.sprites[1].flip_x = s.render.sprites[1].flip_x
-							ns.render.sprites[1].flip_y = s.render.sprites[1].flip_y
-							ns.render.sprites[1].name = s.render.sprites[1].name
-							ns.render.sprites[1].loop = s.render.sprites[1].loop
-							ns.render.sprites[1].ts = s.render.sprites[1].ts
-							ns.render.sprites[1].runs = s.render.sprites[1].runs
-
-							if ne.mercenary then
-								ns.nav_rally.pos = V.vclone(s.nav_rally.pos)
-								ns.nav_rally.center = V.vclone(s.nav_rally.center)
-								ns.nav_rally.new = s.nav_rally.new
-							else
-								ns.nav_rally.pos, ns.nav_rally.center = U.rally_formation_position(i, ne.barrack, ne.barrack.max_soldiers)
-								ns.nav_rally.new = true
-							end
-
-							if ns.melee then
-								for i, a in ipairs(ns.melee.attacks) do
-									if s.melee.attacks[i] then
-										a.ts = s.melee.attacks[i].ts
-									end
-								end
-
-								U.replace_blocker(store, s, ns)
-							end
-
-							ns.soldier.tower_soldier_idx = i
-							ne.barrack.soldiers[i] = ns
-
-							queue_insert(store, ns)
-						end
-
-						s.health.dead = true
-
-						queue_remove(store, s)
-					end
-				end
-			elseif ne.barrack then
-				ne.barrack.rally_pos = V.vclone(ne.tower.default_rally_pos)
-			end
-
-			if ne.tower.type ~= "build_animation" and not ne.tower.hide_dust then
-				local dust = E:create_entity("fx_tower_buy_dust")
-
-				dust.pos.x, dust.pos.y = ne.pos.x, ne.pos.y + 10
-				dust.render.sprites[1].ts = store.tick_ts
-
-				queue_insert(store, dust)
-			end
-
-			if e.tower_upgrade_persistent_data and ne.tower_upgrade_persistent_data then
-				for k, v in pairs(e.tower_upgrade_persistent_data) do
-					if not ne.tower_upgrade_persistent_data[k] then
-						ne.tower_upgrade_persistent_data[k] = v
-					end
-				end
-
-				for _, f in pairs(ne.tower_upgrade_persistent_data.upgrade_functions) do
-					f(ne, store)
-				end
-			end
-		end
-	end
-	perf.stop("tower_upgrade")
-end
-
-sys.game_upgrades = {}
-sys.game_upgrades.name = "game_upgrades"
-
-function sys.game_upgrades:init(store)
-	store.game_upgrades_data = {}
-	store.game_upgrades_data.mage_towers_count = 0
-end
-
-local mage_tower_map = table.to_map(UP.mage_towers)
-
-function sys.game_upgrades:on_insert(entity, store)
-	local mage_bullet_names = UP.mage_tower_bolts
-	local u = UP:get_upgrade("mage_brilliance")
-
-	if entity.tower and u and mage_tower_map[entity.template_name] then
-		local existing_towers = table.filter(store.towers, function(_, e)
-			return mage_tower_map[e.template_name]
-		end)
-		local dps = E:get_template("mod_ray_arcane").dps
-		local bullet_ray_high_elven = E:get_template("ray_high_elven_sentinel").bullet
-		local modifier_pixie = E:get_template("mod_pixie_pickpocket").modifier
-		local f = u.damage_factors[km.clamp(1, #u.damage_factors, #existing_towers + 1)]
-
-		for _, bn in ipairs(mage_bullet_names) do
-			local b = E:get_template(bn).bullet
-
-			if not b._orig_damage_min then
-				b._orig_damage_min = b.damage_min
-				b._orig_damage_max = b.damage_max
-			end
-
-			b.damage_min = ceil(b._orig_damage_min * f)
-			b.damage_max = ceil(b._orig_damage_max * f)
-		end
-
-		if not dps._orig_damage_min then
-			dps._orig_damage_min = dps.damage_min
-			dps._orig_damage_max = dps.damage_max
-		end
-
-		dps.damage_min = ceil(dps._orig_damage_min * f)
-		dps.damage_max = ceil(dps._orig_damage_max * f)
-
-		if not bullet_ray_high_elven._orig_damage_min then
-			bullet_ray_high_elven._orig_damage_min = bullet_ray_high_elven.damage_min
-			bullet_ray_high_elven._orig_damage_max = bullet_ray_high_elven.damage_max
-		end
-
-		bullet_ray_high_elven.damage_min = ceil(bullet_ray_high_elven._orig_damage_min * f)
-		bullet_ray_high_elven.damage_max = ceil(bullet_ray_high_elven._orig_damage_max * f)
-
-		if not modifier_pixie._orig_damage_min then
-			modifier_pixie._orig_damage_min = modifier_pixie.damage_min
-			modifier_pixie._orig_damage_max = modifier_pixie.damage_max
-		end
-
-		modifier_pixie.damage_min = ceil(modifier_pixie._orig_damage_min * f)
-		modifier_pixie.damage_max = ceil(modifier_pixie._orig_damage_max * f)
-
-		local arcane5_disintegrate = E:get_template("tower_arcane_wizard_ray_disintegrate_mod")
-
-		if not arcane5_disintegrate._origin_damage_config then
-			arcane5_disintegrate._origin_damage_config = {}
-			arcane5_disintegrate._origin_damage_config[1] = arcane5_disintegrate.boss_damage_config[1]
-			arcane5_disintegrate._origin_damage_config[2] = arcane5_disintegrate.boss_damage_config[2]
-			arcane5_disintegrate._origin_damage_config[3] = arcane5_disintegrate.boss_damage_config[3]
-		end
-
-		for i = 1, 3 do
-			arcane5_disintegrate.boss_damage_config[i] = ceil(arcane5_disintegrate._origin_damage_config[i] * f)
-		end
-	end
-
-	return true
-end
-
-function sys.game_upgrades:on_remove(entity, store)
-	local mage_towers = UP.mage_towers
-	local mage_bullet_names = UP.mage_tower_bolts
-	local u = UP:get_upgrade("mage_brilliance")
-
-	if entity.tower and u and mage_tower_map[entity.template_name] then
-		local existing_towers = table.filter(store.towers, function(_, e)
-			return mage_tower_map[e.template_name]
-		end)
-		local dps = E:get_template("mod_ray_arcane").dps
-		local bullet_ray_high_elven = E:get_template("ray_high_elven_sentinel").bullet
-		local modifier_pixie = E:get_template("mod_pixie_pickpocket").modifier
-		local f = u.damage_factors[km.clamp(1, #u.damage_factors, #existing_towers - 1)]
-
-		for _, bn in ipairs(mage_bullet_names) do
-			local b = E:get_template(bn).bullet
-
-			b.damage_min = ceil(b._orig_damage_min * f)
-			b.damage_max = ceil(b._orig_damage_max * f)
-		end
-
-		dps.damage_min = ceil(dps._orig_damage_min * f)
-		dps.damage_max = ceil(dps._orig_damage_max * f)
-		bullet_ray_high_elven.damage_min = ceil(bullet_ray_high_elven._orig_damage_min * f)
-		bullet_ray_high_elven.damage_max = ceil(bullet_ray_high_elven._orig_damage_max * f)
-		modifier_pixie.damage_min = ceil(modifier_pixie._orig_damage_min * f)
-		modifier_pixie.damage_max = ceil(modifier_pixie._orig_damage_max * f)
-
-		local arcane5_disintegrate = E:get_template("tower_arcane_wizard_ray_disintegrate_mod")
-
-		for i = 1, 3 do
-			arcane5_disintegrate.boss_damage_config[i] = ceil(arcane5_disintegrate._origin_damage_config[i] * f)
-		end
-	end
-
-	return true
-end
-
-sys.main_script = {}
-sys.main_script.name = "main_script"
-
-function sys.main_script:on_queue(entity, store, insertion)
-	if entity.main_script and entity.main_script.queue then
-		entity.main_script.queue(entity, store, insertion)
-	end
-end
-
-function sys.main_script:on_dequeue(entity, store, insertion)
-	if entity.main_script and entity.main_script.dequeue then
-		entity.main_script.dequeue(entity, store, insertion)
-	end
-end
-
-function sys.main_script:on_insert(entity, store)
-	if entity.main_script and entity.main_script.insert then
-		return entity.main_script.insert(entity, store)
-	else
-		return true
-	end
-end
-
-function sys.main_script:on_update(dt, ts, store)
-	perf.start("main_script")
-	local entities_with_main_script_on_update = store.entities_with_main_script_on_update
-
-	for _, e in pairs(store.entities_with_main_script_on_update) do
-		local s = e.main_script
-
-		if not s.co and s.runs ~= 0 then
-			s.runs = s.runs - 1
-			s.co = coroutine.create(s.update)
-		end
-
-		if s.co then
-			local success, err = coroutine.resume(s.co, e, store)
-
-			-- if coroutine.status(s.co) == "dead" or err ~= nil then
-			--     if err ~= nil then
-			if coroutine.status(s.co) == "dead" or (not success and err ~= nil) then
-				if not success and err ~= nil then
-					-- log.error("Error running coro: %s", debug.traceback(s.co, error))
-					log.error("Error running " .. e.template_name .. " coro: " .. err .. debug.traceback(s.co))
-
-					if LLDEBUGGER then
-						LLDEBUGGER.start()
-					end
-				end
-
-				s.co = nil
-			end
-		end
-	end
-	perf.stop("main_script")
-end
-
-function sys.main_script:on_remove(entity, store)
-	if entity.main_script and entity.main_script.remove then
-		return entity.main_script.remove(entity, store, entity.main_script)
-	else
-		return true
-	end
-end
-
-sys.health = {}
-sys.health.name = "health"
-
-function sys.health:init(store)
-	store.damage_queue = {}
-	store.damages_applied = {}
-end
-
-function sys.health:on_insert(entity, store)
-	if entity.health and not entity.health.hp then
-		entity.health.hp = entity.health.hp_max
-	end
-
-	return true
-end
-
-function sys.health:on_update(dt, ts, store)
-	perf.start("health")
-	local new_damage_queue = {}
-	local damage_queue = store.damage_queue
-	local damages_applied = {}
-	local damages_applied_count = 0
-	local entities = store.entities
-	local damage_queue_len = #damage_queue
-	for i = damage_queue_len, 1, -1 do
-		local d = damage_queue[i]
-		local e = entities[d.target_id]
-
-		if not e then
-		-- block empty
-		else
-			local h = e.health
-
-			if h.dead or band(h.immune_to, d.damage_type) ~= 0 or h.ignore_damage or h.on_damage and not h.on_damage(e, store, d) then
-			-- block empty
-			else
-				local starting_hp = h.hp
-
-				h.last_damage_types = bor(h.last_damage_types, d.damage_type)
-
-				-- 吞噬即死
-				if band(d.damage_type, DAMAGE_EAT) ~= 0 then
-					d.damage_applied = h.hp
-					d.damage_result = bor(d.damage_result, DR_KILL)
-					h.hp = 0
-					damages_applied_count = damages_applied_count + 1
-					damages_applied[damages_applied_count] = d
-				-- 减护甲
-				elseif band(d.damage_type, DAMAGE_ARMOR) ~= 0 then
-					SU.armor_dec(e, d.value)
-
-					d.damage_result = bor(d.damage_result, DR_ARMOR)
-				-- 减法抗
-				elseif band(d.damage_type, DAMAGE_MAGICAL_ARMOR) ~= 0 then
-					SU.magic_armor_dec(e, d.value)
-
-					d.damage_result = bor(d.damage_result, DR_MAGICAL_ARMOR)
-				-- 造成伤害
-				else
-					local actual_damage = U.predict_damage(e, d)
-
-					h.hp = h.hp - actual_damage
-					d.damage_applied = actual_damage
-
-					if starting_hp > 0 and h.hp <= 0 then
-						d.damage_result = bor(d.damage_result, DR_KILL)
-					end
-
-					if actual_damage > 0 then
-						d.damage_result = bor(d.damage_result, DR_DAMAGE)
-
-						if e.regen then
-							e.regen.last_hit_ts = store.tick_ts
-						end
-
-						if d.track_damage then
-							signal.emit("entity-damaged", e, d)
-
-							local source = entities[d.source_id]
-
-							if source and source.track_damage then
-								table.insert(source.track_damage.damaged, {e.id, actual_damage})
-							end
-						end
-					end
-
-					if h.spiked_armor > 0 and e.soldier and d.source_id and e.soldier.target_id == d.source_id then
-						local t = entities[d.source_id]
-
-						if t and t.health and not t.health.dead then
-							local sad = E:create_entity("damage")
-
-							sad.damage_type = DAMAGE_TRUE
-							sad.value = h.spiked_armor * d.value
-							sad.source_id = e.id
-							sad.target_id = t.id
-							new_damage_queue[#new_damage_queue + 1] = sad
-						end
-					end
-
-					damages_applied_count = damages_applied_count + 1
-					damages_applied[damages_applied_count] = d
-				end
-
-				if starting_hp > 0 and h.hp <= 0 then
-					signal.emit("entity-killed", e, d)
-
-					if d.track_kills then
-						local source = entities[d.source_id]
-
-						if source and source.track_kills then
-							table.insert(source.track_kills.killed, e.id)
-						end
-					end
-				end
-			end
-		end
-	end
-
-	local enemies = store.enemies
-	local soldiers = store.soldiers
-
-	for _, e in pairs(enemies) do
-		local h = e.health
-
-		if h.hp <= 0 and not h.dead and not h.ignore_damage then
-			h.hp = 0
-			h.dead = true
-			h.death_ts = store.tick_ts
-			h.delete_after = store.tick_ts + h.dead_lifetime
-
-			if e.health_bar then
-				e.health_bar.hidden = true
-			end
-
-			store.player_gold = store.player_gold + e.enemy.gold
-
-			signal.emit("got-enemy-gold", e, e.enemy.gold)
-		end
-
-		if not h.dead then
-			h.last_damage_types = 0
-		elseif not h.ignore_delete_after and (h.delete_after and store.tick_ts > h.delete_after or h.delete_now) then
-			queue_remove(store, e)
-		end
-	end
-
-	for _, e in pairs(soldiers) do
-		local h = e.health
-
-		if h.hp <= 0 and not h.dead and not h.ignore_damage then
-			h.hp = 0
-			h.dead = true
-			h.death_ts = store.tick_ts
-			h.delete_after = store.tick_ts + h.dead_lifetime
-
-			if e.health_bar then
-				e.health_bar.hidden = true
-			end
-		end
-
-		if not h.dead then
-			h.last_damage_types = 0
-		elseif not e.hero and not h.ignore_delete_after and (h.delete_after and store.tick_ts > h.delete_after or h.delete_now) then
-			queue_remove(store, e)
-		end
-	end
-
-	store.damage_queue = new_damage_queue
-
-	-- 处理伤害结算阶段新加入的伤害，避免遗漏。
-	for i = damage_queue_len + 1, #damage_queue do
-		new_damage_queue[#new_damage_queue + 1] = damage_queue[i]
-	end
-
-	store.damages_applied = damages_applied
-	perf.stop("health")
-end
-
-sys.count_groups = {}
-sys.count_groups.name = "count_groups"
-
-function sys.count_groups:init(store)
-	store.count_groups = {}
-	store.count_groups[COUNT_GROUP_CONCURRENT] = {}
-	store.count_groups[COUNT_GROUP_CUMULATIVE] = {}
-end
-
-function sys.count_groups:on_queue(entity, store, insertion)
-	if insertion and entity.count_group then
-		local c = entity.count_group
-
-		if c.in_limbo then
-			c.in_limbo = nil
-
-			return true
-		end
-
-		local g = store.count_groups
-
-		if not g[c.type][c.name] then
-			g[c.type][c.name] = 0
-		end
-
-		g[c.type][c.name] = g[c.type][c.name] + 1
-
-		signal.emit("count-group-changed", entity, g[c.type][c.name], 1)
-	end
-end
-
-function sys.count_groups:on_dequeue(entity, store, insertion)
-	if insertion then
-		self:on_remove(entity, store)
-	end
-end
-
-function sys.count_groups:on_remove(entity, store)
-	if entity.count_group and not entity.count_group.in_limbo and entity.count_group.type == COUNT_GROUP_CONCURRENT then
-		local c = entity.count_group
-		local g = store.count_groups
-
-		g[c.type][c.name] = km.clamp(0, 1000000000, g[c.type][c.name] - 1)
-
-		signal.emit("count-group-changed", entity, g[c.type][c.name], -1)
-	end
-
-	return true
-end
-
-sys.hero_xp_tracking = {}
-sys.hero_xp_tracking.name = "hero_xp_tracking"
-
-function sys.hero_xp_tracking:on_update(dt, ts, store)
-	perf.start("hero_xp_tracking")
-	for i = 1, #store.damages_applied do
-		local d = store.damages_applied[i]
-
-		if d.xp_gain_factor and d.xp_gain_factor > 0 and d.damage_applied > 0 then
-			local id = d.xp_dest_id or d.source_id
-			local e = store.entities[id]
-
-			if not e or not e.hero then
-			-- block empty
-			else
-				local amount = d.damage_applied * d.xp_gain_factor
-
-				e.hero.xp_queued = e.hero.xp_queued + amount
-			end
-		end
-	end
-	perf.stop("hero_xp_tracking")
-end
-
-sys.pops = {}
-sys.pops.name = "pops"
-
-function sys.pops:on_update(dt, ts, store)
-	perf.start("pops")
-	local damages_applied = store.damages_applied
-	local entities = store.entities
-
-	for i = 1, #damages_applied do
-		local d = damages_applied[i]
-		local pop, target_id = d.pop, d.target_id
-
-		if not pop or not target_id then
-		-- skip
-		else
-			local source = entities[d.source_id]
-			local target = entities[target_id]
-			local pop_entity = (source and (source.enemy or source.soldier)) and source or target
-
-			if not pop_entity then
-				goto continue
-			end
-
-			local pop_chance = d.pop_chance
-			local pop_conds = d.pop_conds
-
-			if (not pop_chance or random() < pop_chance) and (not pop_conds or band(d.damage_result, pop_conds) ~= 0) then
-				local name = pop[random(1, #pop)]
-				local e = E:create_entity(name)
-
-				if e.pop_over_target and target then
-					pop_entity = target
-				end
-
-				local pos_x, pos_y = pop_entity.pos.x, pop_entity.pos.y + e.pop_y_offset
-
-				if pop_entity.unit and pop_entity.unit.pop_offset then
-					pos_y = pos_y + pop_entity.unit.pop_offset.y
-				elseif pop_entity == target and pop_entity.unit and pop_entity.unit.hit_offset then
-					pos_y = pos_y + pop_entity.unit.hit_offset.y
-				end
-
-				e.pos = V.v(pos_x, pos_y)
-				e.render.sprites[1].r = random(-21, 21) * PI / 180
-				e.render.sprites[1].ts = store.tick_ts
-
-				queue_insert(store, e)
-			end
-		end
-
-		::continue::
-	end
-	perf.stop("pops")
-end
-
-sys.timed = {}
-sys.timed.name = "timed"
-
-function sys.timed:on_update(dt, ts, store)
-	perf.start("timed")
-	local entities = store.entities_with_timed
-
-	for _, e in pairs(entities) do
-		local s = e.render.sprites[e.timed.sprite_id]
-
-		if e.timed.disabled then
-		-- block empty
-		elseif s.ts == store.tick_ts then
-		-- block empty
-		elseif e.timed.runs and s.runs == e.timed.runs or e.timed.duration and store.tick_ts - s.ts > e.timed.duration then
-			queue_remove(store, e)
-		end
-	end
-	perf.stop("timed")
-end
-
 sys.tween = {}
 sys.tween.name = "tween"
 
@@ -2276,63 +729,6 @@ function sys.tween:on_update(dt, ts, store)
 	perf.stop("tween_system")
 end
 
-sys.goal_line = {}
-sys.goal_line.name = "goal_line"
-
-function sys.goal_line:on_update(dt, ts, store)
-	perf.start("goal_line")
-	local enemies = store.enemies
-
-	for _, e in pairs(enemies) do
-		local node_index = e.nav_path.ni
-		local end_node = P:get_end_node(e.nav_path.pi)
-
-		if end_node <= node_index and not P.path_connections[e.nav_path.pi] and e.enemy.remove_at_goal_line then
-			signal.emit("enemy-reached-goal", e)
-
-			store.lives = km.clamp(-10000, 10000, store.lives - e.enemy.lives_cost)
-			store.player_gold = store.player_gold + e.enemy.gold
-
-			queue_remove(store, e)
-		end
-	end
-	perf.stop("goal_line")
-end
-
-sys.texts = {}
-sys.texts.name = "texts"
-
-function sys.texts:on_insert(entity, store)
-	if entity.texts then
-		for _, t in pairs(entity.texts.list) do
-			local sprite_id = t.sprite_id
-			local image_name = string.format("text_%s_%s_%s", entity.id, sprite_id, store.tick_ts)
-			local image = F:create_text_image(t.text, t.size, t.alignment, t.font_name, t.font_size, t.color, t.line_height, store.screen_scale, t.fit_height, t.debug_bg)
-
-			I:add_image(image_name, image, "temp_game_texts", store.screen_scale)
-
-			t.image_name = image_name
-			t.image_group = "texts"
-			entity.render.sprites[sprite_id].name = image_name
-			entity.render.sprites[sprite_id].animated = false
-		end
-	end
-
-	return true
-end
-
-function sys.texts:on_remove(entity, store)
-	if entity.texts then
-		for _, t in pairs(entity.texts.list) do
-			if t.image_name then
-				I:remove_image(t.image_name)
-			end
-		end
-	end
-
-	return true
-end
-
 sys.particle_system = {}
 sys.particle_system.name = "particle_system"
 
@@ -2352,292 +748,6 @@ ffi.cdef[[
         int name_idx;
     } particle_t;
 ]]
-
-function sys.particle_system:init(store)
-	self.phase_interp = function(values, phase, default)
-		if not values or #values == 0 then
-			return default
-		end
-
-		if #values == 1 then
-			return values[1]
-		end
-
-		local intervals = #values - 1
-		local interval = floor(phase * intervals)
-		local interval_phase = phase * intervals - interval
-		local a = values[interval + 1]
-		local b = values[interval + 2]
-		local ta = type(a)
-
-		if ta == "table" then
-			local out = {}
-
-			for i = 1, #a do
-				out[i] = a[i] + (b[i] - a[i]) * interval_phase
-			end
-
-			return out
-		elseif ta == "boolean" then
-			return a
-		elseif a ~= nil and b ~= nil then
-			return a + (b - a) * interval_phase
-		end
-
-		return default
-	end
-end
-
-function sys.particle_system:on_insert(entity, store)
-	if entity.particle_system then
-		local ps = entity.particle_system
-
-		ps.emit_ts = (ps.emit_ts and ps.emit_ts or store.tick_ts) + ps.ts_offset
-		ps.ts = store.tick_ts
-		ps.last_pos = {
-			x = 0,
-			y = 0
-		}
-	end
-
-	return true
-end
-
-function sys.particle_system:on_remove(entity, store)
-	if entity.particle_system then
-		local ps = entity.particle_system
-
-		for i = ps.particle_count, 1, -1 do
-			ps.particles[i] = nil
-			ps.frames[i].marked_to_remove = true
-			ps.frames[i] = nil
-		end
-	end
-
-	return true
-end
-
-function sys.particle_system:on_update(dt, ts, store)
-	perf.start("particle_system")
-	local phase_interp = self.phase_interp
-	local particle_systems = store.particle_systems
-
-	for _, e in pairs(particle_systems) do
-		local ps = e.particle_system
-		local e_pos = e.pos
-		local target_rot
-		local particles = ps.particles
-		local frames = ps.frames
-
-		if ps.track_id then
-			local target = store.entities[ps.track_id]
-
-			if target then
-				ps.last_pos.x, ps.last_pos.y = e.pos.x, e.pos.y
-				e_pos.x, e_pos.y = target.pos.x, target.pos.y
-
-				if ps.track_offset then
-					e_pos.x, e_pos.y = e_pos.x + ps.track_offset.x, e_pos.y + ps.track_offset.y
-				end
-
-				if target.render and target.render.sprites[1] then
-					target_rot = target.render.sprites[1].r
-				end
-			else
-				ps.emit = false
-				ps.source_lifetime = 0
-			end
-		end
-
-		if ps.emit_duration and ps.emit then
-			if not ps.emit_duration_ts then
-				ps.emit_duration_ts = ts
-			end
-
-			if ts - ps.emit_duration_ts > ps.emit_duration then
-				ps.emit = false
-			end
-		end
-
-		if not ps.emit then
-			ps.emit_ts = ts + ps.ts_offset
-		elseif ts - ps.emit_ts > 1 / ps.emission_rate then
-			local count = floor((ts - ps.emit_ts) * ps.emission_rate)
-			local particle_lifetime = (ps.particle_lifetime[1] + ps.particle_lifetime[2]) * 0.5
-
-			for i = 1, count do
-				local pts = ps.emit_ts + i / ps.emission_rate
-				ps.particle_count = ps.particle_count + 1
-
-				-- 发生粒子喷射。首先，我们生成粒子，并加入 .particles
-				-- 改用 particle_t 结构体
-				local p = ffi.new("particle_t", 0, 0, ps.emit_rotation and ps.emit_rotation or (ps.track_rotation and target_rot) or (ps.emit_direction + (random() - 0.5) * ps.emit_rotation_spread), 0, 0, ps.spin and random() * (ps.spin[2] - ps.spin[1]) + ps.spin[1] or 0, 1, 1, pts, pts, particle_lifetime, 0)
-
-				particles[ps.particle_count] = p
-
-				local f = {
-					ss = nil,
-					flip_x = false,
-					flip_y = false,
-					pos = {
-						x = 0,
-						y = 0
-					},
-					r = 0,
-					scale = {
-						x = 1,
-						y = 1
-					},
-					anchor = {
-						x = ps.anchor.x,
-						y = ps.anchor.y
-					},
-					offset = {
-						x = 0,
-						y = 0
-					},
-					_draw_order = ps.draw_order and 100000 * ps.draw_order + e.id or floor(pts * 100),
-					z = ps.z,
-					sort_y = ps.sort_y,
-					sort_y_offset = ps.sort_y_offset,
-					alpha = 255,
-					hidden = nil
-				}
-
-				frames[ps.particle_count] = f
-				store.render_frames[#store.render_frames + 1] = f
-
-				if ps.track_id then
-					local factor = (i - 1) / count
-
-					p.pos_x, p.pos_y = ps.last_pos.x + (e_pos.x - ps.last_pos.x) * factor, ps.last_pos.y + (e_pos.y - ps.last_pos.y) * factor
-				else
-					p.pos_x, p.pos_y = e_pos.x, e_pos.y
-				end
-
-				if ps.emit_area_spread then
-					local sp = ps.emit_area_spread
-
-					p.pos_x = p.pos_x + (random() - 0.5) * sp.x * 0.5
-					p.pos_y = p.pos_y + (random() - 0.5) * sp.y * 0.5
-				end
-
-				if ps.emit_offset then
-					p.pos_x = p.pos_x + ps.emit_offset.x
-					p.pos_y = p.pos_y + ps.emit_offset.y
-				end
-
-				if ps.emit_speed then
-					local angle = ps.emission_rate + (random() - 0.5) * ps.emit_spread
-					local len = random() * (ps.emit_speed[2] - ps.emit_speed[1]) + ps.emit_speed[1]
-
-					p.speed_x = cos(angle) * len
-					p.speed_y = sin(angle) * len
-				end
-
-				if ps.scale_var then
-					local factor = random() * (ps.scale_var[2] - ps.scale_var[1]) + ps.scale_var[1]
-
-					p.scale_x = factor
-					p.scale_y = factor
-				-- p.scale_y = ps.scale_same_aspect and factor or random() * factor * 2
-				end
-
-				if ps.names then
-					if ps.cycle_names then
-						if not ps._last_name_idx then
-							ps._last_name_idx = 0
-						end
-
-						ps._last_name_idx = km.zmod(ps._last_name_idx + 1, #ps.names)
-						p.name_idx = ps._last_name_idx
-					else
-						p.name_idx = random(1, #ps.names)
-					end
-				end
-			end
-
-			ps.emit_ts = ps.emit_ts + count * 1 / ps.emission_rate
-		end
-
-		-- 更新 particles，并管理它们的生命周期
-		-- 需要从后往前遍历，以满足 swap 删除方式，同时避免跳过元素
-		for i = ps.particle_count, 1, -1 do
-			do
-				local p = particles[i]
-				local f = frames[i]
-				local phase = (ts - p.ts) / p.lifetime
-
-				if phase >= 1 then
-					-- 不再延迟删除，就地 swap
-					local last_count = ps.particle_count
-
-					particles[i] = particles[last_count]
-					frames[i] = frames[last_count]
-					particles[last_count] = nil
-					frames[last_count] = nil
-					ps.particle_count = last_count - 1
-					f.marked_to_remove = true
-
-					goto label_51_0
-				elseif phase < 0 then
-					phase = 0
-				end
-
-				local tp = ts - p.last_ts
-
-				p.last_ts = ts
-				p.pos_x, p.pos_y = p.pos_x + p.speed_x * tp, p.pos_y + p.speed_y * tp
-				f.pos.x, f.pos.y = p.pos_x, p.pos_y
-				p.r = p.r + p.spin * tp
-				f.r = p.r
-				f.scale.x, f.scale.y = phase_interp(ps.scales_x, phase, 1) * p.scale_x, phase_interp(ps.scales_y, phase, 1) * p.scale_y
-				f.alpha = phase_interp(ps.alphas, phase, 255)
-
-				if ps.sort_y_offsets then
-					f.sort_y_offset = phase_interp(ps.sort_y_offsets, phase, 1)
-				end
-
-				if ps.color then
-					f.color = ps.color
-				end
-
-				local fn
-
-				if ps.animated then
-					local to = ts - p.ts
-
-					if ps.animation_fps then
-						to = to * ps.animation_fps / FPS
-					end
-
-					if p.name_idx > 0 then
-						fn = A:fn(ps.names[p.name_idx], to, ps.loop)
-					else
-						fn = A:fn(ps.name, to, ps.loop)
-					end
-				elseif p.name_idx > 0 then
-					fn = ps.names[p.name_idx]
-				else
-					fn = ps.name
-				end
-
-				f.ss = I:s(fn)
-			end
-
-			::label_51_0::
-		end
-
-		if ps.source_lifetime and ts - ps.ts > ps.source_lifetime then
-			ps.emit = false
-
-			if ps.particle_count == 0 then
-				queue_remove(store, e)
-			end
-		end
-	end
-	perf.stop("particle_system")
-end
 
 sys.render = {}
 sys.render.name = "render"
@@ -3107,557 +1217,40 @@ function sys.render:on_update(dt, ts, store)
 	perf.stop("render")
 end
 
-sys.sound_events = {}
-sys.sound_events.name = "sound_events"
-
-function sys.sound_events:on_insert(entity, store)
-	local se = entity.sound_events
-
-	if se and se.insert then
-		local sounds = se.insert
-
-		if type(sounds) ~= "table" then
-			sounds = {sounds}
-		end
-
-		for _, s in pairs(sounds) do
-			S:queue(s, se.insert_args)
-		end
-	end
-
-	return true
-end
-
-function sys.sound_events:on_remove(entity, store)
-	local se = entity.sound_events
-
-	if se then
-		if se.remove then
-			local sounds = se.remove
-
-			if type(sounds) ~= "table" then
-				sounds = {sounds}
-			end
-
-			for _, s in pairs(sounds) do
-				S:queue(s, se.remove_args)
-			end
-		end
-
-		if se.remove_stop then
-			local sounds = se.remove_stop
-
-			if type(sounds) ~= "table" then
-				sounds = {sounds}
-			end
-
-			for _, s in pairs(sounds) do
-				S:stop(s, se.remove_stop_args)
-			end
-		end
-	end
-
-	return true
-end
-
-sys.seen_tracker = {}
-sys.seen_tracker.name = "seen_tracker"
-
-function sys.seen_tracker:init(store)
-	local slot = storage:load_slot()
-
-	store.seen = slot.seen and slot.seen or {}
-	store.seen_dirty = nil
-end
-
-function sys.seen_tracker:on_insert(entity, store)
-	if (entity.tower or entity.enemy) and not entity.ignore_seen_tracker then
-		U.mark_seen(store, entity.template_name)
-	end
-
-	return true
-end
-
-function sys.seen_tracker:on_update(dt, ts, store)
-	perf.start("seen_tracker")
-	if store.seen_dirty then
-		local slot = storage:load_slot()
-
-		slot.seen = store.seen
-
-		storage:save_slot(slot)
-
-		store.seen_dirty = false
-	end
-	perf.stop("seen_tracker")
-end
-
-sys.editor_overrides = {}
-sys.editor_overrides.name = "editor_overrides"
-
-function sys.editor_overrides:on_insert(entity, store)
-	if not entity.editor then
-		return true
-	end
-
-	local editor = entity.editor
-
-	if editor.components then
-		for _, c in pairs(editor.components) do
-			E:add_comps(entity, c)
-		end
-	end
-
-	if editor.overrides then
-		for k, v in pairs(editor.overrides) do
-			LU.eval_set_prop(entity, k, v)
-		end
-	end
-
-	return true
-end
-
-sys.editor_script = {}
-sys.editor_script.name = "editor_script"
-
-function sys.editor_script:on_insert(entity, store)
-	if entity.editor_script and entity.editor_script.insert then
-		return entity.editor_script.insert(entity, store, entity.editor_script.insert)
-	else
-		return true
-	end
-end
-
-function sys.editor_script:on_remove(entity, store)
-	if entity.editor_script and entity.editor_script.remove then
-		return entity.editor_script.remove(entity, store, entity.editor_script.remove)
-	else
-		return true
-	end
-end
-
-function sys.editor_script:on_update(dt, ts, store)
-	perf.start("editor_script")
-	for _, e in E:filter_iter(store.entities, "editor_script") do
-		local s = e.editor_script
-
-		if not s.update then
-		-- block empty
-		else
-			if not s.co and s.runs ~= 0 then
-				s.runs = s.runs - 1
-				s.co = coroutine.create(s.update)
-			end
-
-			if s.co then
-				local success, error = coroutine.resume(s.co, e, store, s)
-
-				if coroutine.status(s.co) == "dead" or error ~= nil then
-					if error ~= nil then
-						log.error("Error running editor_script coro: %s", debug.traceback(s.co, error))
-					end
-
-					s.co = nil
-				end
-			end
-		end
-	end
-	perf.stop("editor_script")
-end
-
-sys.endless_patch = {}
-sys.endless_patch.name = "endless_patch"
-
-function sys.endless_patch:on_insert(entity, store)
-	if store.level_mode_override == GAME_MODE_ENDLESS then
-		if not entity._endless_strengthened then
-			local endless = store.endless
-
-			entity._endless_strengthened = true
-
-			if entity.enemy then
-				if entity.health.hp_max then
-					entity.health.hp_max = ceil(entity.health.hp_max * store.endless.enemy_health_factor)
-					entity.health.damage_factor = entity.health.damage_factor * store.endless.enemy_health_damage_factor
-					entity.health.instakill_resistance = entity.health.instakill_resistance + store.endless.enemy_instakill_resistance
-				end
-
-				if entity.unit.damage_factor then
-					entity.unit.damage_factor = entity.unit.damage_factor * store.endless.enemy_damage_factor
-				end
-
-				if entity.motion.max_speed then
-					U.speed_mul_self(entity, endless.enemy_speed_factor)
-				end
-
-				entity.enemy.gold = ceil(entity.enemy.gold * store.endless.enemy_gold_factor)
-			elseif entity.soldier then
-				if entity.health and entity.health.hp_max then
-					entity.health.hp_max = ceil(entity.health.hp_max * store.endless.soldier_health_factor)
-					entity.health.hp = entity.health.hp_max
-				-- entity.health.damage_factor = entity.health.damage_factor * store.endless.soldier_health_damage_factor
-				end
-
-				if entity.unit then
-					entity.unit.damage_factor = entity.unit.damage_factor * store.endless.soldier_damage_factor
-					SU.insert_unit_cooldown_buff(store.tick_ts, entity, endless.soldier_cooldown_factor)
-				end
-
-				if entity.hero then
-					entity.unit.damage_factor = entity.unit.damage_factor * store.endless.hero_damage_factor
-
-					SU.insert_unit_cooldown_buff(store.tick_ts, entity, endless.hero_cooldown_factor)
-
-					entity.health.hp_max = ceil(entity.health.hp_max * store.endless.hero_health_factor)
-					entity.health.hp = entity.health.hp_max
-				end
-			elseif entity.tower then
-				SU.insert_tower_damage_factor_buff(entity, endless.tower_damage_factor)
-				SU.insert_tower_cooldown_buff(store.tick_ts, entity, endless.tower_cooldown_factor)
-			end
-		end
-	end
-
-	return true
-end
-
-sys.spatial_index = {}
-sys.spatial_index.name = "spatial_index"
-
-function sys.spatial_index:init(store)
-	package.loaded["spatial_index"] = nil
-	store.enemy_spatial_index = require("spatial_index")
-
-	store.enemy_spatial_index.set_entities(store.enemies)
-	store.enemy_spatial_index.gc_locked(store)
-
-	local seek = require("seek")
-
-	seek.set_id_arrays(store.enemy_spatial_index.get_id_arrays())
-	seek.set_entities(store.enemies)
-end
-
-function sys.spatial_index:on_insert(entity, store)
-	if entity.enemy then
-		store.enemy_spatial_index.insert_entity(entity)
-	end
-
-	return true
-end
-
-function sys.spatial_index:on_remove(entity, store)
-	if entity.enemy then
-		store.enemy_spatial_index.remove_entity(entity)
-	end
-
-	return true
-end
-
-function sys.spatial_index:on_update(dt, ts, store)
-	perf.start("spatial_index")
-	store.enemy_spatial_index.on_update(dt)
-	perf.stop("spatial_index")
-end
-
-sys.last_hook = {}
-sys.last_hook.name = "last_hook"
-
-function sys.last_hook:init(store)
-	store.dead_soldier_count = 0
-	store.enemy_count = 0
-	store.last_hooks = {
-		on_insert = {},
-		on_remove = {}
-	}
-end
-
-function sys.last_hook:on_insert(e, d)
-	if e.enemy then
-		d.enemies[e.id] = e -- 优化分类索引
-
-		if not e.health.patched then
-			if d.level_difficulty == DIFFICULTY_IMPOSSIBLE and d.wave_group_number > 6 then
-				if d.wave_group_number <= 15 then
-					e.health.hp_max = e.health.hp_max * (1 + (d.wave_group_number - 6) * 0.0167)
-				else
-					e.health.hp_max = e.health.hp_max * 1.15
-				end
-			end
-
-			e.health.hp = e.health.hp_max
-			e.health.patched = true
-		end
-
-		if e.enemy.lives_cost == 20 then
-			simulation.store.game_gui:set_boss(e)
-		end
-
-		d.enemy_count = d.enemy_count + 1
-	elseif e.soldier and e.health then
-		d.soldiers[e.id] = e
-	elseif e.modifier then
-		d.modifiers[e.id] = e
-
-		local target = d.entities[e.modifier.target_id]
-
-		if target then
-			if not target._applied_mods then
-				target._applied_mods = {}
-
-				log.error(string.format("！如果看见这条消息，请截下来发给作者 target: %s, mod: %s", target.template_name, e.template_name))
-			end
-
-			local mods = target._applied_mods
-
-			mods[#mods + 1] = e
-		end
-	elseif e.tower then
-		d.towers[e.id] = e
-	elseif e.aura then
-		d.auras[e.id] = e
-	end
-
-	if e.particle_system then
-		d.particle_systems[e.id] = e
-	end
-
-	if e.main_script then
-		if e.main_script.update then
-			d.entities_with_main_script_on_update[e.id] = e
-		end
-	end
-
-	if e.timed then
-		d.entities_with_timed[e.id] = e
-	end
-
-	if e.tween then
-		d.entities_with_tween[e.id] = e
-	end
-
-	if e.render then
-		d.entities_with_render[e.id] = e
-	end
-
-	if e.lights then
-		d.entities_with_lights[e.id] = e
-	end
-
-	if e.ui then
-		d.entities_with_ui[e.id] = e
-	end
-
-	if e.motion and e.motion.max_speed ~= 0 then
-		e.motion.real_speed = e.motion.max_speed
-	end
-
-	for _, hook in pairs(d.last_hooks.on_insert) do
-		hook(e, d)
-	end
-
-	return true
-end
-
-function sys.last_hook:on_remove(e, d)
-	if e.enemy then
-		d.enemies[e.id] = nil -- 优化分类索引
-		d.enemy_count = d.enemy_count - 1
-	elseif e.soldier then
-		d.soldiers[e.id] = nil
-		d.dead_soldier_count = d.dead_soldier_count + 1
-	elseif e.modifier then
-		d.modifiers[e.id] = nil
-
-		local target = d.entities[e.modifier.target_id]
-
-		if target then
-			local mods = target._applied_mods
-
-			if mods then
-				for i = 1, #mods do
-					if mods[i] == e then
-						table.remove(mods, i)
-
-						break
-					end
-				end
-			end
-		end
-	elseif e.tower then
-		d.towers[e.id] = nil
-	elseif e.aura then
-		d.auras[e.id] = nil
-	end
-
-	if e.particle_system then
-		d.particle_systems[e.id] = nil
-	end
-
-	if e.main_script then
-		if e.main_script.update then
-			d.entities_with_main_script_on_update[e.id] = nil
-		end
-	end
-
-	if e.timed then
-		d.entities_with_timed[e.id] = nil
-	end
-
-	if e.tween then
-		d.entities_with_tween[e.id] = nil
-	end
-
-	if e.render then
-		d.entities_with_render[e.id] = nil
-	end
-
-	if e.lights then
-		d.entities_with_lights[e.id] = nil
-	end
-
-	if e.ui then
-		d.entities_with_ui[e.id] = nil
-	end
-
-	for _, hook in pairs(d.last_hooks.on_remove) do
-		hook(e, d)
-	end
-
-	if e._applied_mods then
-		e._applied_mods = nil
-	end
-
-	return true
-end
-
-sys.lights = {}
-sys.lights.name = "lights"
-
-function sys.lights:init(store)
-	store.lights = {}
-end
-
-function sys.lights:on_insert(entity, store)
-	local d = store
-
-	if entity.lights then
-		for i = 1, #entity.lights do
-			local l = entity.lights[i]
-
-			l.pos = {
-				x = entity.pos.x,
-				y = entity.pos.y
-			}
-			d.lights[#d.lights + 1] = l
-		end
-	end
-
-	return true
-end
-
-function sys.lights:on_remove(entity, store)
-	if entity.lights then
-		for i = #entity.lights, 1, -1 do
-			entity.lights[i].marked_to_remove = true
-		-- entity.lights[i] = nil
-		end
-	end
-
-	return true
-end
-
-function sys.lights:on_update(dt, ts, store)
-	perf.start("lights")
-	local d = store
-	local entities = d.entities_with_lights
-	local new_lights = {}
-
-	for _, e in pairs(entities) do
-		for i = 1, #e.lights do
-			local l = e.lights[i]
-
-			if not l.marked_to_remove then
-				l.pos.x, l.pos.y = e.pos.x, e.pos.y
-				new_lights[#new_lights + 1] = l
-			end
-		end
-	end
-
-	if #new_lights > 0 then
-		d.lights = new_lights
-	end
-	perf.stop("lights")
-end
-
--- 美术资源检查模块，在加 assets 参数时启动
-if ASSETS_CHECK_ENABLED then
-	sys.assets_checker = {}
-	sys.assets_checker.name = "assets_checker"
-
-	function sys.assets_checker:init(store)
-		local info_portraits_check_result = {}
-
-		for _, e in pairs(E.entities) do
-			if e.info and e.info.portrait then
-				local s = I:s(e.info.portrait)
-
-				if s == nil then
-					info_portraits_check_result[e.template_name] = e.info.portrait
-				end
-			end
-			if e.timed_attacks and not e.timed_attacks.list[1] then
-				log.error("Entity %s has timed_attacks component but empty list", e.template_name)
-			end
-		end
-
-		local tower_menu_images_check_result = {}
-		local tower_menus_data = require("kr1.data.tower_menus_data")
-
-		for tower_name, tower_menus in pairs(tower_menus_data) do
-			for _, tower_menus_item in pairs(tower_menus) do
-				for _, tower_menus_sub_item in pairs(tower_menus_item) do
-					if tower_menus_sub_item.image then
-						local s = I:s(tower_menus_sub_item.image)
-
-						if s == nil then
-							if not tower_menu_images_check_result[tower_name] then
-								tower_menu_images_check_result[tower_name] = {}
-							end
-
-							tower_menu_images_check_result[tower_name][#tower_menu_images_check_result[tower_name] + 1] = tower_menus_sub_item.image
-						end
-					end
-				end
-			end
-		end
-
-		if next(info_portraits_check_result) ~= nil then
-			log.error("=== info.portrait 资源缺失检查 ===")
-
-			for ename, img in pairs(info_portraits_check_result) do
-				log.error("实体 %s 缺失资源 %s", ename, img)
-			end
-		end
-
-		if next(tower_menu_images_check_result) ~= nil then
-			log.error("=== tower_menus_data 资源缺失检查 ===")
-
-			for tname, imgs in pairs(tower_menu_images_check_result) do
-				for _, img in pairs(imgs) do
-					log.error("实体 %s 缺失资源 %s", tname, img)
-				end
-			end
-		end
-	end
-end
-
-if GEN_WAVES_ENABLED then
-	sys.wave_generator = {}
-	sys.wave_generator.name = "wave_generator"
-
-	function sys.wave_generator:init(store)
-		E:gen_wave(store.level_idx, store.level_mode)
-	end
-end
+SystemsIndex.register_extracted(sys, {
+	log = log,
+	perf = perf,
+	km = km,
+	signal = signal,
+	E = E,
+	UP = UP,
+	U = U,
+	SU = SU,
+	LU = LU,
+	P = P,
+	W = W,
+	GS = GS,
+	DI = DI,
+	S = S,
+	V = V,
+	F = F,
+	I = I,
+	ASSETS_CHECK_ENABLED = ASSETS_CHECK_ENABLED,
+	storage = storage,
+	fts = fts,
+	ceil = ceil,
+	random = random,
+	PI = PI,
+	band = band,
+	bor = bor,
+	queue_insert = queue_insert,
+	queue_remove = queue_remove,
+	floor = floor
+	,
+	ffi = ffi,
+	cos = cos,
+	sin = sin,
+	A = A
+})
 
 return sys
