+++ kr1/boss.lua
@@ -311,7 +311,6 @@ tt.vis.flags = bor(F_ENEMY, F_BOSS)
 tt.pos_castle = vec_2(518, 677)
 tt.souls_aura = "veznan_souls_aura"
 tt.white_circle = "decal_eb_veznan_white_circle"
-tt.taunts.animation = "laught"
 tt.taunts.delay_min = fts(400)
 tt.taunts.delay_max = fts(700)
 tt.taunts.duration = 4
@@ -364,7 +363,7 @@ tt.timed_attacks.list[1].animation = "spellDown"
 tt.timed_attacks.list[1].hit_time = fts(14)
 tt.timed_attacks.list[1].mod = "mod_veznan_tower"
 tt.timed_attacks.list[1].sound = "VeznanHoldCast"
-tt.timed_attacks.list[1].attack_duration = fts(44)
+tt.timed_attacks.list[1].attack_duration = fts(28)
 tt.timed_attacks.list[1].data = {
 	[9] = {13, 2},
 	[10] = {13, 3},
@@ -380,7 +379,7 @@ tt.timed_attacks.list[2].cooldown = 15
 tt.timed_attacks.list[2].hit_time = fts(14)
 tt.timed_attacks.list[2].portal_name = "veznan_portal"
 tt.timed_attacks.list[2].sound = "VeznanPortalSummon"
-tt.timed_attacks.list[2].attack_duration = fts(44)
+tt.timed_attacks.list[2].attack_duration = fts(28)
 tt.timed_attacks.list[2].data = {
 	[6] = {15, 3, {1, 0, 0}},
 	[7] = {10, 2, {1, 0, 0}},
@@ -393,6 +392,24 @@ tt.timed_attacks.list[2].data = {
 	[14] = {15, 3, {1, 1, 1}},
 	[15] = {15, 3, {1, 1, 1}}
 }
+tt.timed_attacks.list[3] = CC("custom_attack")
+tt.timed_attacks.list[3].animation = "soulDrain"
+tt.timed_attacks.list[3].animation_start = "soulDrainStart"
+tt.timed_attacks.list[3].animation_hold = "soulDrainHold"
+tt.timed_attacks.list[3].animation_end = "soulDrainEnd"
+tt.timed_attacks.list[3].cooldown = 5
+tt.timed_attacks.list[3].range = 240
+tt.timed_attacks.list[3].sound = "VeznanSoulDrain"
+tt.timed_attacks.list[3].soul_effect = "veznan_soul"
+tt.timed_attacks.list[3].soul_hand_offset = vec_2(16, 20)
+tt.timed_attacks.list[3].soul_speed = 10 * FPS
+tt.timed_attacks.list[3].heal_hp_factor = 0.05
+tt.timed_attacks.list[3].kill_start_time = fts(9)
+tt.timed_attacks.list[3].kill_end_time = fts(23)
+tt.timed_attacks.list[3].loop_start_time = fts(11)
+tt.timed_attacks.list[3].attack_duration = fts(24)
+tt.timed_attacks.list[3].vis_flags = bor(F_RANGED)
+tt.timed_attacks.list[3].vis_bans = F_NONE
 tt.battle = {
 	ba_animation = "spell",
 	pa_animation = "spell",
